feat: inline commit logic into settings packages
There are 30-100ms delays between each systemd service that runs during the
settings config phase. Avoid some of these delays by running the settings
commit logic as part of the subsequent config phase.
