+++ packages/os/pluto.service
@@ -4,6 +4,8 @@ Description=Generate additional settings for Kubernetes
 # settings generators, and before settings-applier commits all the settings, renders
 # config files, and restarts services.
 After=network-online.target apiserver.service sundog.service
+# Use Before here rather than After=pluto.service in settings-applier.service
+# since this unit is only present on k8s variants.
 Before=settings-applier.service
 Requires=sundog.service
 # We don't want to restart the unit if the network goes offline or apiserver restarts
@@ -15,16 +17,9 @@ RefuseManualStop=true
 
 [Service]
 Type=oneshot
-# pluto needs access to any Kubernetes settings supplied through user-data, along with
-# network-related settings such as proxy servers. Commit any settings that might have
-# been generated during the sundog phase.
-ExecStartPre=/usr/bin/settings-committer
 ExecStart=/usr/bin/pluto
 RemainAfterExit=true
 StandardError=journal+console
 
 [Install]
-# settings-applier requires sundog to succeed as a signal that all settings generators ran
-# successfully. Since pluto is an honorary settings generator, settings-applier also needs
-# it to succeed before it can start.
 RequiredBy=settings-applier.service
