+++ packages/os/sundog.service
@@ -12,8 +12,6 @@ RefuseManualStop=true
 
 [Service]
 Type=oneshot
-# Commit first so we can use settings from user data.
-ExecStartPre=/usr/bin/settings-committer
 ExecStart=/usr/bin/sundog
 RemainAfterExit=true
 StandardError=journal+console
