+++ sources/api/sundog/Cargo.toml
@@ -18,6 +18,7 @@ log.workspace = true
 models.workspace = true
 serde = { workspace = true, features = ["derive"] }
 serde_json.workspace = true
+settings-committer.workspace = true
 simplelog.workspace = true
 shlex.workspace = true
 snafu.workspace = true
