+++ sources/api/thar-be-settings/Cargo.toml
@@ -20,6 +20,7 @@ models.workspace = true
 nix.workspace = true
 schnauzer.workspace = true
 serde_json.workspace = true
+settings-committer.workspace = true
 simplelog.workspace = true
 snafu.workspace = true
 tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
